+++ anyblok_pyramid/config.py
@@ -27,12 +27,6 @@ def get_db_name(request):
     return Configuration.get('db_name')
 
 
-@Configuration.add('preload', label="Preload")
-def define_preload_option(group):
-    group.add_argument('--databases', dest='db_names', nargs="+",
-                       help='List of the database allow to be load')
-
-
 @Configuration.add('wsgi', label="WSGI")
 def define_wsgi_option(group):
     group.add_argument(